+++ packages/types/src/services/consentApi/requests/index.ts
@@ -17,6 +17,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+export * from './ClinicianInvite.js';
 export * from './ConsentRecontact.js';
 export * from './ConsentReleaseData.js';
 export * from './ConsentResearchParticipation.js';
