+++ packages/types/src/services/dataMapper/requests/index.ts
@@ -16,3 +16,7 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+
+export * from './ClinicianInvite.js';
+export * from './ConsentClinicianInvite.js';
+export * from './PIClinicianInvite.js';
