+++ packages/types/src/services/dataMapper/responses/index.ts
@@ -17,5 +17,8 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+export * from './ClinicianInvite.js';
+export * from './ConsentClinicianInvite.js';
 export * from './InformedConsent.js';
+export * from './PIClinicianInvite.js';
 export * from './ParticipantResponse.js';
